Add more StartupSteps to ApplicationStartup (JFR).
Adds a StartupStep for the creation of the target SD module Repository object (e.g. SD JPA). This could vary significantly from 1 SD module to another.

Adds a StartupStep for the creation of the SD Repository Proxy.

Renames the StartupStep wrapping the Repository composition logic to 'spring.data.repository.composition'.

Moves the construction/initialization of the ApplicationStartup instance after the assertions in the getRepository(..) method.